[Style] #25 login api setting
[Style] #25 login api setting
