+++ src/pages/login/index.tsx
@@ -2,7 +2,7 @@ import React, { useState } from 'react';
 import { useRouter } from 'next/router';
 import Head from 'next/head';
 import Image from 'next/image';
-import { loginApi } from 'controllers/services/loginApi';
+import { loginApi } from 'controllers/services/api';
 
 import {
   Link,
